(Wayland) Dedupe VK/GL code
